# pay-guard-backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "Notification_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "patch": {
        "operationId": "Notification_markRead",
        "parameters": [
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "operationId": "Notification_preferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "Notification_updatePreference",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/devices": {
      "post": {
        "operationId": "Notification_registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterNotificationDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/devices/{deviceId}": {
      "delete": {
        "operationId": "Notification_deactivateDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/branches/{branchId}/reconciliations": {
      "post": {
        "operationId": "Reconciliation_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Daily Reconciliations"
        ]
      },
      "get": {
        "operationId": "Reconciliation_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "MATCHED",
                "DISCREPANCY",
                "APPROVED",
                "RETURNED",
                "SUPERSEDED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "settlementAccountId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Daily Reconciliations"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/reconciliations/{reconciliationId}/submit": {
      "post": {
        "operationId": "Reconciliation_submit",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reconciliationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Daily Reconciliations"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/reconciliations/{reconciliationId}/decision": {
      "post": {
        "operationId": "Reconciliation_decide",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reconciliationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Daily Reconciliations"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/reconciliations/{reconciliationId}": {
      "get": {
        "operationId": "Reconciliation_require",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reconciliationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Daily Reconciliations"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayGuard API",
    "description": "Canonical API contract for the PayGuard web portal, Android app, workers, and third-party integrations.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /api/v1/auth/login."
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateNotificationPreferenceDto": {
        "type": "object",
        "properties": {
          "notificationType": {
            "type": "string",
            "enum": [
              "FRAUD_ALERT",
              "CREDIT_ALERT",
              "INCIDENT_ALERT",
              "TRANSACTION_UPDATE",
              "DEVICE_EVENT",
              "FINANCIAL_EVENT",
              "RECONCILIATION_EVENT"
            ]
          },
          "inAppEnabled": {
            "type": "boolean"
          },
          "pushEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "notificationType",
          "inAppEnabled",
          "pushEnabled"
        ]
      },
      "RegisterNotificationDeviceDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web"
            ]
          },
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 4096
          }
        },
        "required": [
          "platform",
          "token"
        ]
      },
...
      "CreateReconciliationDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "settlementAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "reconciliationDate": {
            "type": "string"
          },
          "actualBankBalance": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "differenceExplanation": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          }
        },
        "required": [
          "idempotencyKey",
          "settlementAccountId",
          "reconciliationDate",
          "actualBankBalance",
          "description"
        ]
      },
      "DecideReconciliationDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVED",
              "RETURNED"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "required": [
          "decision",
          "reason"
        ]
      },
...
    }
  }
}